Microservice/deploy-management-server/container-pool/app/requirements.txt
GitPython==3.1.43
fastapi==0.111.1
docker==7.1.0